refactor(security): extract zip-slip guard into shared zipSafety util
Move assertZipEntriesWithin out of McpPackageService into
src/main/utils/zipSafety.ts so backup restore and skill ZIP
install can enforce the same containment check at their extract
boundaries. Message generalized from 'DXT entry' to 'zip entry';
behavior unchanged (tests moved alongside).
