+++ src/main/ai/mcp/__tests__/McpPackageService.test.ts
@@ -11,7 +11,6 @@ vi.mock('@application', async () => {
 const { application } = await import('@application')
 
 const {
-  assertZipEntriesWithin,
   buildResolvedEnv,
   ensurePathWithin,
   McpPackageService,
@@ -92,37 +91,6 @@ describe('ensurePathWithin', () => {
       const target = path.join(baseDir, 'server-name')
       expect(ensurePathWithin(baseDir, target)).toBe(path.resolve(target))
     })
-  })
-})
-
-// C4 (mcp-services-1): node-stream-zip writes each entry at path.join(baseDir, name)
-// with no containment check, so a DXT archive crafted with a `../` entry could
-// overwrite files outside the extraction dir. The guard must reject before extraction.
-describe('assertZipEntriesWithin', () => {
-  const baseDir = path.join('/', 'tmp', 'dxt_extract')
-
-  // Skip on Windows — path separator / resolution semantics differ.
-  const testFn = process.platform === 'win32' ? it.skip : it
-
-  testFn('accepts entries nested inside the base dir', () => {
-    expect(() => assertZipEntriesWithin(['manifest.json', 'server/index.js', 'assets/icon.png'], baseDir)).not.toThrow()
-  })
-
-  testFn('rejects a parent-traversal entry', () => {
-    expect(() => assertZipEntriesWithin(['../escape.txt'], baseDir)).toThrow('zip-slip')
-  })
-
-  testFn('rejects a deep traversal entry mixed with safe entries', () => {
-    expect(() => assertZipEntriesWithin(['manifest.json', '../../../etc/cron.d/evil'], baseDir)).toThrow('zip-slip')
-  })
-
-  testFn('rejects an absolute-path entry', () => {
-    expect(() => assertZipEntriesWithin(['/etc/passwd'], baseDir)).toThrow('zip-slip')
-  })
-
-  testFn('rejects a sibling-prefix entry that is not actually nested', () => {
-    // `/tmp/dxt_extract_evil` shares the base string prefix but is not within the base dir.
-    expect(() => assertZipEntriesWithin(['../dxt_extract_evil/x'], baseDir)).toThrow('zip-slip')
   })
 })
